# docs/swagger.yaml
basePath: /
definitions:
  error.errorResponse:
    properties:
      message:
        type: string
    type: object
  error.errorResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  model.ResponseSuccess:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  model.User:
    properties:
      bio:
        default: I am Golang dev
        type: string
      city:
        default: Navoi
        type: string
      email:
        default: phapp0224mb@gmail.com
        type: string
      first_name:
        default: MRB
        type: string
      interests:
        items:
          type: string
        type: array
      nickname:
        default: mrb
        type: string
      password:
        default: 0224mb
        type: string
      phone:
        default: +9989 93 753 65 71
        type: string
      second_name:
        default: HERO
        type: string
    type: object
host: https://mediumuz.herokuapp.com
info:
  contact:
    email: phapp0224mb@gmail.com
    name: Bakhodir Yashin Mansur
  description: API Server for MediumuZ Application
  termsOfService: https://github.com/BakhodiribnYashinibnMansur/Medium_uz
  title: MediumuZ API
  version: "1.0"
paths:
  /api/account/follower:
    get:
      consumes:
      - application/json
      description: Follower user
      operationId: follower-account
      parameters:
      - description: Param ID
        in: query
        name: followerID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add Follower
      tags:
      - Account
  /api/account/following:
    get:
      consumes:
      - application/json
      description: Following user
      operationId: following-account
      parameters:
      - description: Param ID
        in: query
        name: followingID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add Following
      tags:
      - Account
  /api/account/get:
    get:
      consumes:
      - application/json
      description: return account data. if you send id = "" or null return current
        user data. if send id = number return number user data .
      operationId: get-account
      parameters:
      - description: id
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Account Data
      tags:
      - Account
  /api/account/get-followers:
    get:
      consumes:
      - application/json
      description: Follower user
      operationId: get-follower-account
      parameters:
      - description: 'Offset '
        in: query
        name: offset
        type: integer
      - description: 'Limit '
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Account Follower
      tags:
      - Account
  /api/account/get-followings:
    get:
      consumes:
      - application/json
      description: Followings user
      operationId: get-following-account
      parameters:
      - description: 'Offset '
        in: query
        name: offset
        type: integer
      - description: 'Limit '
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Account Following
      tags:
      - Account
...
  /api/account/sendcode:
    get:
      consumes:
      - application/json
      description: send code to email for  verification
      operationId: send-code-email
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Send code for  Verification Email
      tags:
      - Account
  /api/account/update:
    put:
      consumes:
      - application/json
      description: update account give data
      operationId: update-account
      parameters:
      - description: account info
        in: body
        name: input
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Account
      tags:
      - Account
  /api/account/upload-image:
    patch:
      consumes:
      - application/json
      - multipart/form-data
      description: Upload Account Image
      operationId: upload-image
      parameters:
      - description: file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      - application/octet-stream
      - image/png
      - image/jpeg
      - image/jpg
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Upload Account Image
      tags:
      - Account
  /api/account/verify:
    get:
      consumes:
      - application/json
      description: verification email with code
      operationId: verify-email
      parameters:
      - description: code
        in: query
        name: code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/error.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/error.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/error.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Verification Email
      tags:
      - Account
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"